+++ examples/test/sizeof/test_sizeof_bad_array.pas
@@ -1,0 +1,6 @@
+program TestSizeofBadArray;
+var
+    arr: array[1..5] of integer;
+begin
+    writeln(sizeOf(arr));
+end.