+++ examples/test/sizeof/test_sizeof_bad_record.pas
@@ -1,0 +1,8 @@
+program TestSizeofBadRecord;
+type
+    TRec = record
+        name: string;
+    end;
+begin
+    writeln(sizeOf(TRec));
+end.