+++ examples/test/sizeof/test_sizeof_bad_scalarvar.pas
@@ -1,0 +1,6 @@
+program TestSizeofBadScalarvar;
+var
+    b: byte;
+begin
+    writeln(sizeOf(b));
+end.