+++ examples/test/sizeof/test_sizeof_bad_string.pas
@@ -1,0 +1,4 @@
+program TestSizeofBadString;
+begin
+    writeln(sizeOf(string));
+end.