+++ examples/test/sizeof/test_sizeof_bad_unknown.pas
@@ -1,0 +1,4 @@
+program TestSizeofBadUnknown;
+begin
+    writeln(sizeOf(notDeclaredAtAll));
+end.