+++ examples/test/sizeof/test_sizeof_record_var.pas
@@ -1,0 +1,20 @@
+program TestSizeofRecordVar;
+type
+    TRec = record
+        flag: byte;
+        total: integer;
+    end;
+var
+    globalR: TRec;
+
+procedure CheckLocal;
+var
+    localR: TRec;
+begin
+    writeln(sizeOf(localR));
+end;
+
+begin
+    writeln(sizeOf(globalR));
+    CheckLocal;
+end.