+++ examples/test/sizeof/test_sizeof_scalars.pas
@@ -1,0 +1,9 @@
+program TestSizeofScalars;
+begin
+    writeln(sizeOf(integer));
+    writeln(sizeOf(real));
+    writeln(sizeOf(boolean));
+    writeln(sizeOf(byte));
+    writeln(sizeOf(shortint));
+    writeln(sizeOf(word));
+end.